adding liability release form to competition, verifying students, etc

+++ api/routes/competitions.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, jsonify, Response, request, current_app
+from flask import Blueprint, jsonify, Response, request, current_app, url_for
 from typing import Dict, Optional, Tuple
 import http_status_codes as status
 from pymongo.errors import WriteError, OperationFailure
@@ -6,6 +6,7 @@ from datetime import date, datetime
 from pydantic import ValidationError
 from bson.objectid import ObjectId
 import logging
+import gridfs
 from models import CreateCompetitionRequest, GetCompetitionResponse
 
 competitions_blueprint = Blueprint("competitions", __name__)
@@ -19,12 +20,24 @@ db_competitions_collection = current_app.config['DB_COMPETITION_COLLECTION']
 @competitions_blueprint.route('/competitions/create', methods=["POST"])
 def create_competition() -> Tuple[Response, int]:
     try:
-        create_competition_request: CreateCompetitionRequest = CreateCompetitionRequest.model_validate_json(request.data)
+        create_competition_request: CreateCompetitionRequest = CreateCompetitionRequest.model_validate_json(request.form.get('competition'))
 
         create_competition_dict: Dict = create_competition_request.model_dump()
         create_competition_dict['created_at'] = datetime.now()
         db = client[db_name]
         collection = db[db_competitions_collection]
+        liability_release_form_file_id = None
+        fs = gridfs.GridFS(db)
+        
+        # save file if present
+        if 'liability_release_form_file' in request.files:
+            file = request.files['liability_release_form_file']
+            liability_release_form_file_id = fs.put(file, filename=file.filename)
+        else:
+            return jsonify({"error": "Cannot create competition without liability release form"}), status.BAD_REQUEST
+
+        create_competition_dict['liability_release_form_file_id'] = liability_release_form_file_id
+
         response = collection.insert_one(create_competition_dict)
 
         if response.inserted_id is not None:
@@ -53,7 +66,6 @@ def create_competition() -> Tuple[Response, int]:
 @competitions_blueprint.route('/competitions/get')
 def get_competitions() -> Tuple[Response, int]:
     try:
-
         db = client[db_name]
         collection = db[db_competitions_collection]
 
@@ -65,7 +77,8 @@ def get_competitions() -> Tuple[Response, int]:
                     "competition_name": document["competition_name"],
                     "created_at": document["created_at"],
                     "registration_deadline": document["registration_deadline"],
-                    "is_active": document["is_active"]
+                    "is_active": document["is_active"],
+                    "liability_release_form": url_for('files.download_file', file_id=document["liability_release_form_file_id"], _external=True),
                 }
                 validated_competition: GetCompetitionResponse = GetCompetitionResponse.model_validate(competition)
                 competition_dict = validated_competition.model_dump()
@@ -146,7 +159,8 @@ def get_competition_details():
             "competition_name": document["competition_name"],
             "created_at": document["created_at"],
             "registration_deadline": document["registration_deadline"],
-            "is_active": document["is_active"]
+            "is_active": document["is_active"],
+            "liability_release_form": url_for('files.download_file', file_id=document["liability_release_form_file_id"], _external=True),
         }
 
         validated_competition: GetCompetitionResponse = GetCompetitionResponse.model_validate(competition)
@@ -179,8 +193,14 @@ def update_or_delete_competition(competition_id) -> Tuple[Response, int]:
         
         db = client[db_name]
         collection = db[db_competitions_collection]
+        fs = gridfs.GridFS(db)
+        competition = collection.find_one({"_id": ObjectId(competition_id)})
+        if competition is None:
+            return jsonify({"error":"Could not find any competition with that competition_id"}), status.BAD_REQUEST
 
         if request.method == "DELETE":
+            # delete liability release form
+            fs.delete(competition["liability_release_form_file_id"])
             delete_attempt = collection.delete_one({"_id": ObjectId(competition_id)})
 
             if delete_attempt.deleted_count == 1:
@@ -190,8 +210,22 @@ def update_or_delete_competition(competition_id) -> Tuple[Response, int]:
                 return jsonify({"error": "Failed to delete competition"}), status.INTERNAL_SERVER_ERROR
 
         elif request.method == "PUT":
-            update_competition_request: CreateCompetitionRequest = CreateCompetitionRequest.model_validate_json(request.data)
+            update_competition_request: CreateCompetitionRequest = CreateCompetitionRequest.model_validate_json(request.form.get("competition"))
             update_competition_data: Dict = update_competition_request.model_dump()
+
+
+            if request.form.get("delete_old_liability_release_form") == "true":
+                if 'liability_release_form_file' in request.files:
+                    # delete old liability release form
+                    fs.delete(competition["liability_release_form_file_id"])
+
+                    file = request.files['liability_release_form_file']
+                    liability_release_form_file_id = fs.put(file, filename=file.filename)
+                    update_competition_data['liability_release_form_file_id'] = liability_release_form_file_id
+                else:
+                    return jsonify({"error": "Cannot have competition without liability release form"}), status.BAD_REQUEST
+
+
             update_attempt = collection.update_one(
                     {"_id": ObjectId(competition_id)},
                     {"$set": update_competition_data}
